+++ dessia_common/workflow.py
@@ -876,8 +876,12 @@ class Export(Block):
         Block.__init__(self, inputs=[TypedVariable(type_=method_type.class_)], outputs=[output], name=name)
 
     def evaluate(self, values):
-        res = getattr(values[self.inputs[0]], self.method_type.name)()
-        return [res]
+        if self.text:
+            stream = StringFile()
+        else:
+            stream = BinaryFile()
+        getattr(values[self.inputs[0]], self.method_type.name)(stream)
+        return [stream]
 
     def _export_format(self, block_index: int):
         args = {"block_index": block_index}
@@ -893,6 +897,7 @@ class ExportJson(Export):
         if not export_name:
             self.export_name += "_json"
         self.extension = "json"
+        self.text = True
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
         dict_ = Block.to_dict(self)
@@ -916,6 +921,7 @@ class ExportExcel(Export):
         if not export_name:
             self.export_name += "_xlsx"
         self.extension = "xlsx"
+        self.test = False
 
     def to_dict(self, use_pointers=True, memo=None, path: str = '#'):
         dict_ = Block.to_dict(self)
@@ -2189,10 +2195,20 @@ class WorkflowState(DessiaObject):
         else:
             output_value = None
 
-        values = {workflow.variables[int(i)]: deserialize(v,
-                                                          global_dict=dict_,
-                                                          pointers_memo=pointers_memo)
-                  for i, v in dict_['values'].items()}
+        values = {}
+        if 'values' in dict_:
+            for i, value in dict_['values'].items():
+                values[workflow.variables[int(i)]] = deserialize(value,
+                                                                 global_dict=dict_,
+                                                                 pointers_memo=pointers_memo,
+                                                                 path=f'{path}/values/{i}')
+        elif 'variable_values' in dict_:
+            # Retrocompat with variable value may be removed after v0.10.0:
+            for i, value in dict_['variable_values'].items():
+                values[workflow.variables[int(i)]] = deserialize(value,
+                                                                 global_dict=dict_,
+                                                                 pointers_memo=pointers_memo,
+                                                                 path=f'{path}/variable_values/{i}')
 
         input_values = {int(i): deserialize(v, global_dict=dict_, pointers_memo=pointers_memo)
                         for i, v in dict_['input_values'].items()}
@@ -2310,7 +2326,7 @@ class WorkflowState(DessiaObject):
 
         evaluated_blocks = []
         something_activated = True
-        while something_activated and (self.progress < 1 or export):
+        while something_activated:  # and (self.progress < 1 or export)
             something_activated = False
 
             for pipe in self._activable_pipes():
@@ -2504,15 +2520,17 @@ class WorkflowRun(WorkflowState):
             end_time = time.time()
         self.end_time = end_time
         self.execution_time = end_time - start_time
+        self.variable_values = {workflow.variable_indices(k): v for k, v in values.items() if k.memorize}
         WorkflowState.__init__(self, workflow=workflow, input_values=input_values, activated_items=activated_items,
                                values=values, start_time=start_time, output_value=output_value, log=log, name=name)
 
-    @property
-    def variable_values(self):
+    def to_dict(self, use_pointers: bool = True, memo=None, path: str = '#'):
         """
-        Jsonable saved values
+        Adds variable values to super WorkflowState dict
         """
-        return {self.workflow.variable_indices(k): v for k, v in self.values.items() if k.memorize}
+        dict_ = WorkflowState.to_dict(self, use_pointers=use_pointers, memo=memo, path=path)
+        dict_["variable_values"] = {str(k): serialize(v) for k, v in self.variable_values.items()}
+        return dict_
 
     def display_settings(self) -> List[DisplaySetting]:
         """